openvswitch: support setting OpenFlow datapath ID
By default, Open vSwitch will generate the OpenFlow datapath ID of a
bridge based on the MAC address of one of its ports. Due to this, it's
possible that the datapath ID changes when new ports are added. When the
datapath ID changes, Open vSwitch disconnects from the controller, as
there is no way to notify the controller that the datapath ID has
changed.

Add an option to set the datapath ID so that the above situation can be
avoided. The option either takes 16 hex characters, or when prefixed
with @, the name of a UCI interface section. In the latter case, it will
lookup the MAC address of the interface, and prefix it with 6f77, which
is hex for "ow".

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>